server/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
asyncpg>=0.30,<1
python-dotenv>=1.1,<2
httpx>=0.28,<1
pillow>=11.0,<12
watchdog>=6.0,<7
sse-starlette>=2.0,<3
PyJWT>=2.10,<3
bcrypt>=4.3,<5
pydantic>=2.10,<3
